refactoring page and grid components
