+++ components/Grid.tsx
@@ -1,12 +1,14 @@
 "use client";
 
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import { Note, Interval, SynthType } from "@/types";
 import { NOTES } from "@/constants";
 import { usePreferences } from "@/contexts/PreferencesContext";
 import { getNoteHighlight } from "@/utils/NoteHighlighter";
 import { PianoSamplerSingleton } from "@/utils/PianoSamplerSingleton";
 import * as Tone from "tone";
+import { useInstrument } from "@/hooks/useInstrument";
+import LoadingSpinner from "@/components/LoadingSpinner";
 
 interface GridProps {
   pattern?: Interval;
@@ -26,79 +28,7 @@ const Grid: React.FC<GridProps> = ({
 }) => {
   const { preferences } = usePreferences();
   const activeSynthType = propsSynthType ?? preferences.synthType;
-  const [instrument, setInstrument] = useState<
-    Tone.Synth | Tone.Sampler | Tone.PolySynth | null
-  >(null);
-  const [isLoading, setIsLoading] = useState(false);
-
-  useEffect(() => {
-    let newInstrument: typeof instrument = null;
-    let mounted = true;
-
-    const setupInstrument = async () => {
-      setIsLoading(true);
-
-      if (activeSynthType === "piano") {
-        newInstrument = await PianoSamplerSingleton.getInstance();
-      } else {
-        newInstrument = new Tone.PolySynth(Tone.Synth, {
-          oscillator: {
-            type: "fatsawtooth",
-          },
-          envelope: {
-            attack: 0.03,
-            decay: 0.3,
-            sustain: 0.4,
-            release: 1.5,
-          },
-        })
-          .chain(
-            // Add chorus for width and richness
-            new Tone.Chorus({
-              frequency: 2.5,
-              delayTime: 3.5,
-              depth: 0.7,
-              wet: 0.3,
-            }),
-            // Add reverb for space and depth
-            new Tone.Reverb({
-              decay: 2,
-              wet: 0.2,
-            }),
-            // Add a subtle compression to glue it together
-            new Tone.Compressor({
-              threshold: -24,
-              ratio: 3,
-              attack: 0.03,
-              release: 0.25,
-            }),
-            // EQ to shape the tone
-            new Tone.EQ3({
-              low: 2,
-              mid: 0,
-              high: 1,
-              lowFrequency: 250,
-              highFrequency: 2500,
-            }),
-          )
-          .toDestination();
-      }
-
-      if (mounted) {
-        setInstrument(newInstrument);
-        setIsLoading(false);
-      }
-    };
-
-    setupInstrument();
-
-    return () => {
-      mounted = false;
-      if (newInstrument && activeSynthType !== "piano") {
-        newInstrument.dispose();
-      }
-    };
-  }, [activeSynthType]);
+  const { instrument, isLoading } = useInstrument(activeSynthType);
 
   // Preload piano samples when component mounts
   useEffect(() => {
@@ -136,33 +66,7 @@ const Grid: React.FC<GridProps> = ({
   };
 
   if (isLoading) {
-    return (
-      <div className="flex justify-center items-center min-h-[200px] w-full">
-        <div className="flex items-center gap-2 text-lg font-medium text-gray-700 dark:text-gray-300">
-          <svg
-            className="animate-spin h-5 w-5"
-            xmlns="http://www.w3.org/2000/svg"
-            fill="none"
-            viewBox="0 0 24 24"
-          >
-            <circle
-              className="opacity-25"
-              cx="12"
-              cy="12"
-              r="10"
-              stroke="currentColor"
-              strokeWidth="4"
-            ></circle>
-            <path
-              className="opacity-75"
-              fill="currentColor"
-              d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-            ></path>
-          </svg>
-          Loading synthesizer...
-        </div>
-      </div>
-    );
+    return <LoadingSpinner />;
   }
 
   // Use props.visible if provided, otherwise check preferences
